Fix session teardown: avoid broken ActivityResult path, fix int/long bug
SessionActivity has documentLaunchMode="always", which per Android's
documented startActivityForResult constraints means it cannot return a
result to the caller - the original ActivityResultLauncher-based design
would have silently never fired, leaking the SSH tunnel and never
restoring the wallpaper after every session.

Also found or the PARAM_INSTANCE extra was read via bundle.getInt() while
the actual instance handles are native pointers returned as long
(LibFreeRDP.newInstance/freerdp_new) - unreachable/untested upstream
since nothing in the tree currently sets that extra, but would silently
truncate a 64-bit pointer and miss the GlobalApp session map lookup,
NPEing on session.getSurface(). Fixed to bundle.getLong().

New design: FreeRdpLauncher creates the FreeRDP session directly via
GlobalApp.createSession() (in-memory, no bookmark DB needed at all -
dropped the Room/ManualBookmarkGateway approach entirely) and launches
SessionActivity with that session's instance via PARAM_INSTANCE.
MainActivity registers a GlobalApp.SessionEventListener on that instance
before starting the activity and tears down the SSH tunnel from
onConnectionFailure/onDisconnected, which reflects the actual FreeRDP
session lifecycle instead of guessing at Activity result semantics.

+++ client/Android/Studio/freeRDPCore/src/main/java/com/freerdp/freerdpcore/presentation/SessionActivity.java
@@ -474,7 +474,9 @@ public class SessionActivity extends AppCompatActivity
 		}
 		else if (bundle.containsKey(PARAM_INSTANCE))
 		{
-			int inst = bundle.getInt(PARAM_INSTANCE);
+			// Instance handles are native pointers (LibFreeRDP.newInstance/freerdp_new),
+			// which don't fit in 32 bits on a 64-bit process - must read as long, not int.
+			long inst = bundle.getLong(PARAM_INSTANCE);
 			session = GlobalApp.getSession(inst);
 			bitmap = session.getSurface().getBitmap();
 			bindSession();
